+++ docs/CHANGELOG.md
@@ -1,3 +1,4 @@
+- 2022-03-02  feat: timer coundown
 - 2022-03-02  feat: tkinter main window styling 8f0fd59
 - 2022-03-01  feat: initial project scaffolding 1c54bb7
 - 2022-02-28  Initial commit 5262834